% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_components.R
\name{county_components}
\alias{county_components}
\title{Create a ploty chart showing the components of change by county}
\usage{
county_components(input, output, session, fips = 0, name = "Colorado")
}
\description{
This function will generate a UI element for shiny that
}

