% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_componentsUI.R
\name{county_componentsUI}
\alias{county_componentsUI}
\title{Create a ploty chart showing the components of change by county}
\usage{
county_componentsUI(id)
}
\description{
This function will generate a UI element for shiny that
}

