% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoEncode.R
\name{geoEncode}
\alias{geoEncode}
\title{Create a drop-down menu of county or municipality names}
\usage{
geoEncode(name, type = "county")
}
\arguments{
\item{name}{geography name output from geoInput}

\item{type}{county is default, "place" will return place names}
}
\description{
This function will generate a UI element for shiny that
}

