% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoInput.R
\name{geoInput}
\alias{geoInput}
\title{Create a drop-down menu of county or municipality names}
\usage{
geoInput(type = "county")
}
\arguments{
\item{type}{county is default, "place" will return place names}
}
\description{
This function will generate a UI element for shiny that
}

